+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/ClientConfigDiagnosticsTest.java
@@ -167,7 +167,7 @@ public class ClientConfigDiagnosticsTest {
         assertThat(objectNode.get("id").asInt()).isEqualTo(1);
         assertThat(objectNode.get("machineId").asText()).isEqualTo(machineId);
         assertThat(objectNode.get("numberOfClients").asInt()).isEqualTo(2);
-        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, mm: false, prgns: [null])");
+        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, readConsistencyStrategy: null,  mm: false, prgns: [null])");
         assertThat(objectNode.get("connCfg").get("rntbd").asText()).isEqualTo("null");
         assertThat(objectNode.get("connCfg").get("gw").asText()).isEqualTo("null");
         assertThat(objectNode.get("connCfg").get("other").asText()).isEqualTo("(ed: false, cs: false, rv: true)");
@@ -200,7 +200,7 @@ public class ClientConfigDiagnosticsTest {
         assertThat(objectNode.get("id").asInt()).isEqualTo(1);
         assertThat(objectNode.get("machineId").asText()).isEqualTo(machineId);
         assertThat(objectNode.get("numberOfClients").asInt()).isEqualTo(2);
-        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, mm: false, prgns: [null])");
+        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, readConsistencyStrategy: null,  mm: false, prgns: [null])");
         assertThat(objectNode.get("connCfg").get("rntbd").asText()).isEqualTo("(cto:PT5S, nrto:PT5S, icto:PT0S, ieto:PT1H, mcpe:130, mrpc:30, cer:true)");
 
         String http2Enabled = Configs.isHttp2Enabled() ? "true" : "false";
@@ -237,7 +237,7 @@ public class ClientConfigDiagnosticsTest {
         assertThat(objectNode.get("id").asInt()).isEqualTo(1);
         assertThat(objectNode.get("machineId").asText()).isEqualTo(machineId);
         assertThat(objectNode.get("numberOfClients").asInt()).isEqualTo(2);
-        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, mm: false, prgns: [null])");
+        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, readConsistencyStrategy: null,  mm: false, prgns: [null])");
         assertThat(objectNode.get("connCfg").get("rntbd").asText()).isEqualTo("null");
 
         String http2Enabled = Configs.isHttp2Enabled() ? "true" : "false";
@@ -311,7 +311,7 @@ public class ClientConfigDiagnosticsTest {
         assertThat(objectNode.get("id").asInt()).isEqualTo(1);
         assertThat(objectNode.get("machineId").asText()).isEqualTo(machineId);
         assertThat(objectNode.get("numberOfClients").asInt()).isEqualTo(2);
-        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, mm: false, prgns: [westus1,westus2])");
+        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, readConsistencyStrategy: null,  mm: false, prgns: [westus1,westus2])");
         assertThat(objectNode.get("connCfg").get("rntbd").asText()).isEqualTo("null");
 
         String http2Enabled = Configs.isHttp2Enabled() ? "true" : "false";
@@ -364,7 +364,7 @@ public class ClientConfigDiagnosticsTest {
         assertThat(objectNode.get("id").asInt()).isEqualTo(1);
         assertThat(objectNode.get("machineId").asText()).isEqualTo(machineId);
         assertThat(objectNode.get("numberOfClients").asInt()).isEqualTo(2);
-        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, mm: false, prgns: [null])");
+        assertThat(objectNode.get("consistencyCfg").asText()).isEqualTo("(consistency: null, readConsistencyStrategy: null,  mm: false, prgns: [null])");
         assertThat(objectNode.get("connCfg").get("rntbd").asText()).isEqualTo("null");
         assertThat(objectNode.get("connCfg").get("gw").asText()).isEqualTo("null");
         assertThat(objectNode.get("connCfg").get("other").asText()).isEqualTo("(ed: false, cs: false, rv: true)");