+++ sdk/eventgrid/azure-messaging-eventgrid-namespaces/src/main/java/com/azure/messaging/eventgrid/namespaces/EventGridReceiverAsyncClient.java
@@ -142,16 +142,7 @@ public final class EventGridReceiverAsyncClient {
      *          (Required){
      *             lockToken: String (Required)
      *             error (Required): {
-     *                 code: String (Required)
-     *                 message: String (Required)
-     *                 target: String (Optional)
-     *                 details (Optional): [
-     *                     (recursive schema, see above)
-     *                 ]
-     *                 innererror (Optional): {
-     *                     code: String (Optional)
-     *                     innererror (Optional): (recursive schema, see innererror above)
-     *                 }
+     *                 error (Required): (recursive schema, see error above)
      *             }
      *         }
      *     ]
@@ -214,16 +205,7 @@ public final class EventGridReceiverAsyncClient {
      *          (Required){
      *             lockToken: String (Required)
      *             error (Required): {
-     *                 code: String (Required)
-     *                 message: String (Required)
-     *                 target: String (Optional)
-     *                 details (Optional): [
-     *                     (recursive schema, see above)
-     *                 ]
-     *                 innererror (Optional): {
-     *                     code: String (Optional)
-     *                     innererror (Optional): (recursive schema, see innererror above)
-     *                 }
+     *                 error (Required): (recursive schema, see error above)
      *             }
      *         }
      *     ]
@@ -277,16 +259,7 @@ public final class EventGridReceiverAsyncClient {
      *          (Required){
      *             lockToken: String (Required)
      *             error (Required): {
-     *                 code: String (Required)
-     *                 message: String (Required)
-     *                 target: String (Optional)
-     *                 details (Optional): [
-     *                     (recursive schema, see above)
-     *                 ]
-     *                 innererror (Optional): {
-     *                     code: String (Optional)
-     *                     innererror (Optional): (recursive schema, see innererror above)
-     *                 }
+     *                 error (Required): (recursive schema, see error above)
      *             }
      *         }
      *     ]
@@ -340,16 +313,7 @@ public final class EventGridReceiverAsyncClient {
      *          (Required){
      *             lockToken: String (Required)
      *             error (Required): {
-     *                 code: String (Required)
-     *                 message: String (Required)
-     *                 target: String (Optional)
-     *                 details (Optional): [
-     *                     (recursive schema, see above)
-     *                 ]
-     *                 innererror (Optional): {
-     *                     code: String (Optional)
-     *                     innererror (Optional): (recursive schema, see innererror above)
-     *                 }
+     *                 error (Required): (recursive schema, see error above)
      *             }
      *         }
      *     ]